backend/requirements.txt
dataclasses==0.6
dataclasses-json==0.6.4
marshmallow==3.20.2
mypy-extensions==1.0.0
packaging==23.2
pex==2.2.1
PyYAML==6.0.1
setproctitle==1.3.3
setuptools==69.1.1
typing-inspect==0.9.0
typing_extensions==4.10.0
